Native: don't treat any compiler test task as up-to-date
There are number of problems with up-to-date checks in tests, leading to
a test not being rebuilt or rerun when a developer expects it to.

Things became worse after updating to Gradle 7+:
test run tasks are considered up-to-date if the test data
has been changed, even after clean.

Workaround this by forcing all tasks in the
:kotlin-native:backend.native:tests project to be not up-to-date.